requirements.txt
streamlit==1.28.0
pandas==2.1.1
numpy==1.24.3
plotly==5.17.0
scikit-learn==1.3.1
xgboost==2.0.0
lightgbm==4.1.0
catboost==1.2.2

# RAG Engine dependencies
langchain>=0.2.0
langchain-community>=0.2.0
langchain-core>=0.2.0
langchain-huggingface>=0.0.3
langchain-google-genai>=1.0.0
langchain-anthropic>=0.1.0
langchain-openai>=0.1.0
faiss-cpu>=1.7.4
sentence-transformers>=2.2.0
python-dotenv>=1.0.0
